# go/docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: name0euph@gmail.com
    name: name0euph
  description: This is a simple todo API.
  title: Todo API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: 正常性を確認するためのエンドポイント
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 正常性エンドポイント
      tags:
      - others
  /login:
    post:
      consumes:
      - application/json
      description: 既存のユーザーでログインする
      parameters:
      - description: ユーザー名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: パスワード
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: ログイン
      tags:
      - users
  /logout:
    post:
      consumes:
      - application/json
      description: ユーザがログアウトする
      parameters:
      - description: ユーザー名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: パスワード
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: ログアウト
      tags:
      - users
  /signup:
    post:
      consumes:
      - application/json
      description: 新しいユーザーを作成する
      parameters:
      - description: ユーザー名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: パスワード
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: サインアップ
      tags:
      - users
swagger: "2.0"